requirements.txt
streamlit==1.28.2
ultralytics==8.0.216
rasterio==1.3.9
Pillow==10.1.0
rasterio==1.3.9
black==24.4.0